Close streams on app pause, re-connect on resume, move multi-sig subscription code
