+++ src/subscriptions/account-data.ts
@@ -1,7 +1,7 @@
 import { Horizon, Server } from "stellar-sdk"
 import { trackError } from "../context/notifications"
 import { loadAccount, waitForAccountData } from "../lib/account"
-import { createStreamDebouncer, trackStreamError } from "../lib/stream"
+import { createStreamDebouncer, manageStreamConnection, trackStreamError } from "../lib/stream"
 import { createSubscriptionTarget, SubscriptionTarget } from "../lib/subscription"
 
 export interface ObservedAccountData {
@@ -35,28 +35,29 @@ export function createAccountDataSubscription(
   const { debounceError, debounceMessage } = createStreamDebouncer<Server.AccountRecord>()
   const { propagateUpdate, subscriptionTarget } = createSubscriptionTarget(createEmptyAccountData(accountPubKey))
 
-  const subscribeToStream = (cursor: string = "now") => {
-    horizon
-      .accounts()
-      .accountId(accountPubKey)
-      .cursor(cursor)
-      .stream({
-        reconnectTimeout: 8000,
-        onmessage(accountData: Server.AccountRecord) {
-          debounceMessage(accountData, () => {
-            propagateUpdate({
-              ...subscriptionTarget.getLatest(),
-              ...(accountData as any)
+  const subscribeToStream = (cursor: string = "now") =>
+    manageStreamConnection(() => {
+      return horizon
+        .accounts()
+        .accountId(accountPubKey)
+        .cursor(cursor)
+        .stream({
+          reconnectTimeout: 8000,
+          onmessage(accountData: Server.AccountRecord) {
+            debounceMessage(accountData, () => {
+              propagateUpdate({
+                ...subscriptionTarget.getLatest(),
+                ...(accountData as any)
+              })
             })
-          })
-        },
-        onerror(error: any) {
-          debounceError(error, () => {
-            trackStreamError(new Error("Account data update stream errored."))
-          })
-        }
-      } as any)
-  }
+          },
+          onerror(error: any) {
+            debounceError(error, () => {
+              trackStreamError(new Error("Account data update stream errored."))
+            })
+          }
+        } as any)
+    })
 
   const setup = async () => {
     const initialAccountData = await loadAccount(horizon, accountPubKey)
